refactor: 기타 컴포넌트 type 지정 및 리팩토링
