+++ src/index.tsx
@@ -17,7 +17,9 @@ Sentry.init({
 if (process.env.REACT_APP_GOOGLE_ANALYTICS_TRACKING_ID) {
   ReactGA.initialize(process.env.REACT_APP_GOOGLE_ANALYTICS_TRACKING_ID);
 }
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement,
+);
 root.render(
   <React.StrictMode>
     <App />
